% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_bairisch}
\alias{get_bairisch}
\title{get_bairisch}
\usage{
get_bairisch(txt)
}
\arguments{
\item{txt}{provide a text you want to translate from German into Bavarian}
}
\description{
Translates German to Bavarian
}
