% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{schwab_symbolDetail}
\alias{schwab_symbolDetail}
\title{Get ticker details}
\usage{
schwab_symbolDetail(tickers = c("AAPL", "SPY"), accessTokenList = NULL)
}
\arguments{
\item{tickers}{valid ticker(s) or symbol(s)}

\item{accessTokenList}{A valid Access Token must be set using the output from
\code{\link{schwab_auth3_accessToken}}. The most recent Access Token will be
used by default unless one is manually passed into the function.}
}
\value{
data frame of ticker details
}
\description{
Get identifiers and fundamental data for a specific ticker
}
\examples{
\dontrun{

# Details for Apple
schwab_symbolDetail('AAPL')

}
}
