% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{schwab_userPreferences}
\alias{schwab_userPreferences}
\title{Get User Preferences}
\usage{
schwab_userPreferences(accessTokenList = NULL)
}
\arguments{
\item{accessTokenList}{A valid Access Token must be set using the output from
\code{\link{schwab_auth3_accessToken}}. The most recent Access Token will be
used by default unless one is manually passed into the function.}
}
\value{
Dataframe output of account details
}
\description{
Returns a dataframe output for account preferences associated with user
}
\examples{
\dontrun{

# Access Token must be set using schwab_auth3_accessToken
# Market hours for the current date
schwab_userPreferences()

}
}
