% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schwabr.R
\docType{package}
\name{schwabr-package}
\alias{schwabr}
\alias{schwabr-package}
\title{schwabr: A Package for Using the Schwab Trade API}
\description{
The schwabr package provides four categories of important functions:
\enumerate{
    \item Authenticating into a TD Brokerage account
    \item Working with orders
    \item Getting account information
    \item Pulling pricing/market data
}
}
\details{
Details for the functions can be found within the function help pages.
Detailed examples are also available in the README.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://altanalytics.github.io/schwabr/}
  \item Report bugs at \url{https://github.com/altanalytics/schwabr/issues}
}

}
\author{
\strong{Maintainer}: Anthony Trevisan \email{anthonytrevisan@gmail.com}

}
\keyword{internal}
