% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_schwab}
\alias{get_schwab}
\title{get_schwab}
\usage{
get_schwab(txt)
}
\arguments{
\item{txt}{provide a text you want to translate from German into Swabian (a maximum of 1009 characters per call)}
}
\description{
Translates German to Swabian
}
